Code update with directory -- nats-test 'Quick Refactor'
