+++ nats-test/src/listener.ts
@@ -1,5 +1,6 @@
-import nats, { Message, Stan } from 'node-nats-streaming';
+import nats  from 'node-nats-streaming';
 import { randomBytes } from 'crypto';
+import { TicketCreatedListener } from './events/ticket-created-listener';
 
 console.clear();
 
@@ -25,59 +26,3 @@ process.on('SIGINT', () => stan.close());
 process.on('SIGTERM', () => stan.close());
 
 /** Adding Handlers END */
-
-abstract class Listener {
-    abstract subject: string;
-    abstract queueGroupName: string;
-    abstract onMessage(data: any, msg: Message): void;
-    private client: Stan;
-    protected ackWait = 5 * 1000;
-
-    constructor(client: Stan) {
-        this.client = client;
-    }
-
-    subscriptionOptions(){
-        return this.client
-            .subscriptionOptions()
-            .setDeliverAllAvailable()
-            .setManualAckMode(true)
-            .setAckWait(this.ackWait)
-            .setDurableName(this.queueGroupName);
-    }
-
-    listen(){
-        const subscription = this.client.subscribe(
-            this.subject,
-            this.queueGroupName,
-            this.subscriptionOptions()
-        );
-
-        subscription.on('message', (msg: Message) =>{
-            console.log(
-                `Message received: ${this.subject} / ${this.queueGroupName}`
-            );
-
-            const parsedData = this.parseMessage(msg);
-            this.onMessage(parsedData, msg);
-        });
-    }
-
-    parseMessage(msg: Message){
-        const data = msg.getData();
-        return typeof data === 'string'
-        ? JSON.parse(data)
-        : JSON.parse(data.toString('utf8'));
-    }
-}
-
-class TicketCreatedListener extends Listener {
-    subject = 'ticket:created';
-    queueGroupName = 'payments-service';
-
-    onMessage(data: any, msg: Message){
-        console.log('Event data!', data);
-
-        msg.ack();
-    }
-}
